refactor(iCGM Sensor): Separated Sensor from Generator
